Minor Updates
Works up to 0.5 now, still some issue with using 0.6 where both client
and server lock up, client is "waiting" and server is "waiting. occurs
when the server is waiting for the EOT and the client is waiting for all
packets to be acked. Need better way of determining cumulative acks.